Move SharedLogLogic-specific getCheckedOperations from superclass, move operation resolving to AbstractLogLogic
